Use quota in a dictionary
