+++ fastapi_server/src/rio_app/components/replay_pack_builder/models.py
@@ -1,4 +1,5 @@
 # pyright: reportUnknownMemberType=false, reportUnknownArgumentType=false, reportAttributeAccessIssue=false, reportImplicitOverride=false
+from collections import deque
 from dataclasses import dataclass
 from hashlib import md5
 from pathlib import Path
@@ -9,6 +10,23 @@ from rio import FileInfo
 
 REPLAYS_FOLDER = Path(__file__).parents[4] / "data" / "replay_pack_builder"
 REPLAYS_FOLDER.mkdir(parents=True, exist_ok=True)
+
+# Once a file size threshold exceeds, start deleting files
+FILES_IN_ORDER = deque[Path](p for p in REPLAYS_FOLDER.glob("**/*.SC2Replay"))
+
+quota = {
+    "QUOTA_LIMIT": 10 * 2**30,  # 10 gigabyte of replays can be uploaded in total
+    "quota_used": sum(p.stat().st_size for p in FILES_IN_ORDER),
+}
+
+
+# On server start: Delete all uploading files
+def delete_pending_upload_files():
+    for f in REPLAYS_FOLDER.glob("*.uploading"):
+        f.unlink(missing_ok=True)
+
+
+delete_pending_upload_files()
 
 
 @dataclass
@@ -76,6 +94,7 @@ class ReplayFile(BaseModel):
         replay_path.parent.mkdir(parents=True, exist_ok=True)
         _ = temp_replay_path.write_bytes(data)
         _ = temp_replay_path.rename(replay_path)
+        FILES_IN_ORDER.append(replay_path)
         self.path = replay_path
         return self.size
